requirements.txt
pyvoip
vosk
python-dotenv
